Use long for doctor IDs

+++ app/src/main/java/com/example/symptotrack/Doc/DetallePaciente.java
@@ -66,7 +66,7 @@ public class DetallePaciente extends AppCompatActivity {
             Toast.makeText(this, "Debes iniciar sesión como doctor", Toast.LENGTH_SHORT).show();
             return;
         }
-        int doctorId = (int) session.getId();
+        long doctorId = session.getId();
 
         api.patientDetail(doctorId, patientId).enqueue(new Callback<ApiResponse<PatientDetailDto>>() {
             @Override public void onResponse(Call<ApiResponse<PatientDetailDto>> call, Response<ApiResponse<PatientDetailDto>> resp) {
