+++ app/src/main/java/com/example/symptotrack/doctor/DoctorItem.java
@@ -1,11 +1,11 @@
 package com.example.symptotrack.doctor;
 
 public class DoctorItem {
-    public final int id;
+    public final long id;
     public final String nombreCompleto;
     public final String email;
 
-    public DoctorItem(int id, String nombreCompleto, String email) {
+    public DoctorItem(long id, String nombreCompleto, String email) {
         this.id = id;
         this.nombreCompleto = nombreCompleto;
         this.email = email;
