+++ app/src/main/java/com/example/symptotrack/net/dto/DoctorDto.java
@@ -1,7 +1,7 @@
 package com.example.symptotrack.net.dto;
 
 public class DoctorDto {
-    public int doctor_id;
+    public long doctor_id;
     public String first_name;
     public String last_name;
     public String email;
