+++ app/src/main/java/com/example/symptotrack/net/dto/LoginData.java
@@ -2,7 +2,7 @@ package com.example.symptotrack.net.dto;
 
 public class LoginData {
     public String role; // "user" o "doctor"
-    public int id;      // id de user o doctor_id
+    public long id;      // id de user o doctor_id
     public String first_name;
     public String last_name;
 }
