+++ app/src/main/java/com/example/symptotrack/net/dto/ShareRequest.java
@@ -1,12 +1,12 @@
 package com.example.symptotrack.net.dto;
 
 public class ShareRequest {
-    public int doctor_id;
+    public long doctor_id;
     public long patient_id;
     public String note;   // opcional
     public String fecha;  // "yyyy-MM-dd" opcional
 
-    public ShareRequest(int doctor_id, long patient_id, String note, String fecha) {
+    public ShareRequest(long doctor_id, long patient_id, String note, String fecha) {
         this.doctor_id = doctor_id;
         this.patient_id = patient_id;
         this.note = note;
