+++ app/src/main/java/com/example/symptotrack/net/dto/ShareResponse.java
@@ -2,7 +2,7 @@ package com.example.symptotrack.net.dto;
 
 public class ShareResponse {
     public long id;
-    public int doctor_id;
+    public long doctor_id;
     public long patient_id;
     public String fecha;
 }
